[Feature]Refactore code to improve maintainability
