+++ app/api/v2/views/auth_view.py
@@ -10,7 +10,7 @@ from passlib.hash import sha256_crypt
 from flask_restful import Resource, reqparse
 
 # import class products
-from app.api.v2.models.store_model import Users
+from app.api.v2.models.auth_model import Users
 from app.api.v2.utils.authorization import admin_required
 blacklist = set()
